split status sensor out of sensor.py, fix test harness
