+++ custom_components/simple_timer/sensor.py
@@ -23,26 +23,18 @@ from homeassistant.helpers.event import (
     async_track_point_in_utc_time,
     async_track_time_interval,
 )
-from homeassistant.helpers.dispatcher import (
-    async_dispatcher_connect,
-    async_dispatcher_send,
-)
+from homeassistant.helpers.dispatcher import async_dispatcher_send
 from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.helpers.storage import Store
 from homeassistant.util import dt as dt_util
 
-from homeassistant.helpers import device_registry as dr, entity_registry as er
+from homeassistant.helpers import entity_registry as er
 from homeassistant.helpers.device_registry import DeviceInfo
 
 from .const import (
     DOMAIN,
     WARNING_MSG_OFFLINE,
     SIGNAL_STATE_UPDATED,
-    STATUS_IDLE,
-    STATUS_ACTIVE,
-    STATUS_DELAYED_START,
-    STATUS_SCHEDULED,
-    STATUS_OPTIONS,
     EVENT_TIMER_STARTED,
     EVENT_TIMER_EXTENDED,
     EVENT_TIMER_CANCELLED,
@@ -50,6 +42,8 @@ from .const import (
     EVENT_SCHEDULE_SET,
     EVENT_SCHEDULE_CANCELLED,
 )
+from .helpers import device_info_for_switch
+from .status_sensor import TimerStatusSensor
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -101,52 +95,6 @@ def duration_to_seconds(duration: float, unit: str) -> float:
     if unit in ["d", "day", "days"]:
         return duration * 86400
     return duration * 60  # minutes is the default unit
-
-
-def derive_timer_status(timer_state: str, reverse_mode: bool, has_schedule: bool) -> str:
-    """Map the runtime sensor's internal flags to a single user-facing status.
-
-    A schedule can stay armed while a timer is running (a repeating schedule
-    arms its next fire immediately), so a running timer deliberately wins.
-
-    Kept as a module-level pure function so it can be unit tested without
-    standing up an entity.
-    """
-    if timer_state == "active":
-        return STATUS_DELAYED_START if reverse_mode else STATUS_ACTIVE
-    if has_schedule:
-        return STATUS_SCHEDULED
-    return STATUS_IDLE
-
-
-def device_info_for_switch(hass: HomeAssistant, switch_entity_id: str | None) -> DeviceInfo | None:
-    """Return DeviceInfo that groups an entity onto the switch's device.
-
-    Reuses the switch device's identifiers so HA merges our entities into that
-    device rather than creating a second one. Shared by both sensors.
-    """
-    if not switch_entity_id:
-        return None
-
-    # Access the Entity Registry to find the registry entry for the switch
-    ent_reg = er.async_get(hass)
-    entity_entry = ent_reg.async_get(switch_entity_id)
-
-    # If the switch doesn't exist or isn't linked to a device, we can't link
-    if not entity_entry or not entity_entry.device_id:
-        return None
-
-    # Access the Device Registry to get the device details
-    dev_reg = dr.async_get(hass)
-    device_entry = dev_reg.async_get(entity_entry.device_id)
-
-    if not device_entry:
-        return None
-
-    return DeviceInfo(
-        connections=device_entry.connections,
-        identifiers=device_entry.identifiers,
-    )
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry, async_add_entities) -> None:
@@ -314,6 +262,24 @@ class TimerRuntimeSensor(SensorEntity, RestoreEntity):
         """
         super().async_write_ha_state()
         async_dispatcher_send(self.hass, SIGNAL_STATE_UPDATED.format(self._entry_id))
+
+    @property
+    def timer_state(self) -> str:
+        """Internal timer state: "idle" or "active"."""
+        return self._timer_state
+
+    @property
+    def timer_reverse_mode(self) -> bool:
+        """Whether the current (or most recent) timer is a delayed start.
+
+        Sticky once set - see the note in __init__.
+        """
+        return self._timer_reverse_mode
+
+    @property
+    def has_armed_schedule(self) -> bool:
+        """Whether a scheduled start is currently armed."""
+        return self._scheduled_fire_at is not None
 
     @property
     def status_entity_id(self) -> str | None:
@@ -2416,114 +2382,3 @@ class TimerRuntimeSensor(SensorEntity, RestoreEntity):
         
         _LOGGER.info(f"Simple Timer: [{self._entry_id}] Daily usage reset: {old_state}s -> 0s")
 
-
-class TimerStatusSensor(SensorEntity):
-    """Non-numeric companion sensor exposing the timer's status as its state.
-
-    Exists so the timer shows up in HA's logbook and device Activity feed. The
-    runtime sensor cannot: it carries a unit_of_measurement, and logbook skips
-    continuous sensors to avoid drowning in numbers.
-
-    Deliberately exposes NO extra state attributes. The card locates a timer
-    instance by scanning every sensor.* for one carrying both `entry_id` and
-    `switch_entity_id` and taking the first match (_determineEffectiveEntities
-    in timer-card.ts). Adding those attributes here would let this entity win
-    that race and break existing cards, including bundles already shipped to
-    users who will not rebuild. Do not add them.
-    """
-
-    _attr_has_entity_name = False
-    _attr_should_poll = False
-    _attr_device_class = SensorDeviceClass.ENUM
-    _attr_options = STATUS_OPTIONS
-
-    def __init__(self, hass: HomeAssistant, entry: ConfigEntry):
-        """Initialize the status sensor."""
-        self.hass = hass
-        self._entry = entry
-        self._entry_id = entry.entry_id
-        self._entry_id_short = self._entry_id[:8]
-        self._switch_entity_id = entry.data.get("switch_entity_id")
-
-        self._attr_unique_id = f"timer_status_{self._entry_id}"
-        self._attr_native_value = STATUS_IDLE
-
-        self._unsub_dispatcher = None
-
-    @property
-    def device_info(self) -> DeviceInfo | None:
-        """Group with the switch's device, same as the runtime sensor."""
-        return device_info_for_switch(self.hass, self._switch_entity_id)
-
-    @property
-    def instance_title(self) -> str:
-        """Current instance title, mirroring the runtime sensor."""
-        if self._entry.title:
-            return self._entry.title
-        return self._entry.data.get("name") or "Timer"
-
-    @property
-    def name(self) -> str:
-        """Return the name of the sensor."""
-        return f"{self.instance_title} Status ({self._entry_id_short})"
-
-    @property
-    def icon(self) -> str:
-        """Icon reflecting the current status."""
-        if self._attr_native_value == STATUS_ACTIVE:
-            return "mdi:timer-play"
-        if self._attr_native_value == STATUS_DELAYED_START:
-            return "mdi:timer-sand"
-        if self._attr_native_value == STATUS_SCHEDULED:
-            return "mdi:timer-cog"
-        return "mdi:timer-outline"
-
-    def _read_runtime_sensor(self) -> "TimerRuntimeSensor | None":
-        """Return this entry's runtime sensor, if it is loaded."""
-        entry_data = self.hass.data.get(DOMAIN, {}).get(self._entry_id)
-        if not isinstance(entry_data, dict):
-            return None
-        return entry_data.get("sensor")
-
-    @callback
-    def _handle_state_updated(self) -> None:
-        """Recompute status; write only when it actually changed.
-
-        The signal fires on every runtime sensor write, which is once a second
-        while a timer runs. Filtering here keeps the recorder to a handful of
-        rows per timer instead of one per second.
-        """
-        runtime = self._read_runtime_sensor()
-        if runtime is None:
-            return
-
-        new_status = derive_timer_status(
-            runtime._timer_state,
-            getattr(runtime, "_timer_reverse_mode", False),
-            bool(runtime._scheduled_fire_at),
-        )
-
-        if new_status != self._attr_native_value:
-            self._attr_native_value = new_status
-            self.async_write_ha_state()
-
-    async def async_added_to_hass(self) -> None:
-        """Subscribe to runtime sensor updates and seed the initial status."""
-        await super().async_added_to_hass()
-
-        self._unsub_dispatcher = async_dispatcher_connect(
-            self.hass,
-            SIGNAL_STATE_UPDATED.format(self._entry_id),
-            self._handle_state_updated,
-        )
-
-        # Seed from the runtime sensor if it is already loaded, so a restart
-        # mid-timer does not report idle until the next write.
-        self._handle_state_updated()
-
-    async def async_will_remove_from_hass(self) -> None:
-        """Drop the dispatcher subscription."""
-        if self._unsub_dispatcher:
-            self._unsub_dispatcher()
-            self._unsub_dispatcher = None
-        await super().async_will_remove_from_hass()
